requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2
langchain>=0.0.267
langchain-openai>=0.0.2
langchain-core>=0.1.0
langgraph>=0.0.16
openai>=1.0.0
python-dotenv>=0.19.0 